made things flow better and fixed recursion bugs

+++ bot.py
@@ -5,14 +5,6 @@ import logging
 import discord
 import asyncio
 
-#read dictionary
-words = []
-dictin = open("dictionary", "a+")
-dictin.seek(0)
-dictionary = dictin.readlines()
-for i in range(0, len(dictionary)):
-    words.append(dictionary[i].rstrip())
-    
 #logging
 logger = logging.getLogger("discord")
 logger.setLevel(logging.DEBUG)
@@ -44,6 +36,14 @@ async def on_ready():
     print("Logged in as " + str(client.user.name) + "(" + str(client.user.id) + ")")
     print("------")
 
+#read dictionary
+words = []
+dictin = open("dictionary", "a+")
+dictin.seek(0)
+dictionary = dictin.readlines()
+for i in range(0, len(dictionary)):
+    words.append(dictionary[i].rstrip())
+
 #message recieved
 @client.event
 async def on_message(message):
@@ -51,7 +51,7 @@ async def on_message(message):
     if(str(message.channel)=="simple_word_room"):
         text = str(message.content).translate(dict.fromkeys(map(ord, string.punctuation))).split()
         for i in range(0, len(text)):
-            if not text[i] in words:
+            if message.author != client.user and not text[i] in words:
                 await client.delete_message(message)
                 await client.send_message(message.channel, "please speak simple words")
                 break
